+++ src/core/deferred.ts
@@ -19,14 +19,14 @@ export class Deferred<T = Object, K = Object> {
      * @param {*=} opt_data
      * @return {undefined}
      */
-    resolve(opt_data?: T | T[]): void {
+    resolve(opt_data?: T): void {
         this._promise.resolve(opt_data);
     }
     /**
      * @param {*=} opt_data
      * @return {undefined}
      */
-    reject(opt_data?: K | K[]): void {
+    reject(opt_data?: K): void {
         this._promise.reject(opt_data);
     }
 }